requirements.txt
kafka-python==2.0.2
pytest>=7.0.0
mock>=4.0.0
typing>=3.7.4
python-json-logger>=2.0.0
dataclasses>=0.6; python_version < '3.7'
docker-compose>=1.29.2